[refactor] test code 위치 변경
