+++ test/unit/feeds.service.spec.ts
@@ -1,5 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { FeedsService } from './feeds.service';
+// import { FeedsService } from './feeds.service';
+import { FeedsService } from '../../src/feeds/feeds.service';
 
 describe('FeedsService', () => {
   let service: FeedsService;
